refactor: interval, timer, and user services
Refactor interval, timer, and user services to handle empty or invalid
data, and to improve error reporting. This ensures more robust and
user-friendly data handling.

ever-co/ever-gauzy-desktop#44

+++ packages/desktop-libs/src/lib/offline/services/interval.service.ts
@@ -19,15 +19,20 @@ export class IntervalService implements IIntervalService<IntervalTO> {
 		this._timerService = new TimerService();
 		this._offlineMode = DesktopOfflineModeHandler.instance;
 	}
+
 	public async create(interval: IntervalTO): Promise<void> {
 		try {
+			if (!interval) {
+				return console.error('WARN[INTERVAL_SERVICE]: Without data interval cannot be created');
+			}
 			interval.activities = JSON.stringify(interval.activities);
 			interval.screenshots = JSON.stringify(interval.screenshots) as any;
 			await this._intervalDAO.save(interval);
 		} catch (error) {
-			throw new AppError('INTERVALSRVCE', error);
+			throw new AppError('[INTERVAL_SERVICE]', error);
 		}
 	}
+
 	public async backedUpNoSynced(): Promise<IntervalTO[]> {
 		try {
 			const user = await this._userService.retrieve();
@@ -41,31 +46,33 @@ export class IntervalService implements IIntervalService<IntervalTO> {
 			return [];
 		}
 	}
+
 	public async destroy(interval: Partial<IntervalTO>): Promise<void> {
 		try {
+			if (!interval) {
+				return console.error('WARN[INTERVAL_SERVICE]: Without data interval cannot be destroyed');
+			}
 			await this._intervalDAO.delete(interval);
 		} catch (error) {
-			throw new AppError('INTERVALSRVCE', error);
+			throw new AppError('[INTERVAL_SERVICE]', error);
 		}
 	}
+
 	public async synced(interval: IntervalTO): Promise<void> {
 		try {
+			if (!interval) {
+				return console.error('WARN[INTERVAL_SERVICE]: Without data interval cannot be synced');
+			}
 			const intervalToUpdate = new Interval(interval);
 			intervalToUpdate.synced = true;
-			intervalToUpdate.activities = JSON.stringify(
-				intervalToUpdate.activities
-			);
-			intervalToUpdate.screenshots = JSON.stringify(
-				intervalToUpdate.screenshots
-			) as any;
-			await this._intervalDAO.update(
-				intervalToUpdate.id,
-				intervalToUpdate.toObject()
-			);
+			intervalToUpdate.activities = JSON.stringify(intervalToUpdate.activities);
+			intervalToUpdate.screenshots = JSON.stringify(intervalToUpdate.screenshots) as any;
+			await this._intervalDAO.update(intervalToUpdate.id, intervalToUpdate.toObject());
 		} catch (error) {
-			throw new AppError('INTERVALSRVCE', error);
+			throw new AppError('[INTERVAL_SERVICE]', error);
 		}
 	}
+
 	public async backedUpAllNoSynced(): Promise<IntervalTO[]> {
 		try {
 			const user = await this._userService.retrieve();
@@ -75,6 +82,7 @@ export class IntervalService implements IIntervalService<IntervalTO> {
 			return [];
 		}
 	}
+
 	public async countNoSynced(): Promise<number> {
 		try {
 			const user = await this._userService.retrieve();
@@ -86,6 +94,7 @@ export class IntervalService implements IIntervalService<IntervalTO> {
 			return 0;
 		}
 	}
+
 	public async screenshots(): Promise<any[]> {
 		try {
 			const user = await this._userService.retrieve();
@@ -96,24 +105,19 @@ export class IntervalService implements IIntervalService<IntervalTO> {
 		}
 	}
 
-	public async removeIdlesTime(
-		startedAt: Date,
-		stoppedAt: Date
-	): Promise<string[]> {
+	public async removeIdlesTime(startedAt: Date, stoppedAt: Date): Promise<string[]> {
 		try {
+			if (!startedAt || !stoppedAt) {
+				console.error(
+					`WARN[INTERVAL_SERVICE]: Without startedAt or stoppedAt dates, we cannot remove idles time`
+				);
+				return [];
+			}
 			const user = await this._userService.retrieve();
 			let remoteIds = [];
 			this._offlineMode.enabled
-				? await this._intervalDAO.deleteLocallyIdlesTime(
-					startedAt,
-					stoppedAt,
-					user
-				)
-				: (remoteIds = await this._intervalDAO.deleteIdlesTime(
-					startedAt,
-					stoppedAt,
-					user
-				));
+				? await this._intervalDAO.deleteLocallyIdlesTime(startedAt, stoppedAt, user)
+				: (remoteIds = await this._intervalDAO.deleteIdlesTime(startedAt, stoppedAt, user));
 			return remoteIds.map(({ remoteId }) => remoteId);
 		} catch (error) {
 			console.error(error);
@@ -127,17 +131,23 @@ export class IntervalService implements IIntervalService<IntervalTO> {
 	 */
 	public async remove(id: number): Promise<void> {
 		try {
+			if (!id || typeof id !== 'number') {
+				return console.error('WARN[INTERVAL_SERVICE]: No interval id, cannot remove');
+			}
 			await this._intervalDAO.delete({ id });
 		} catch (error) {
-			throw new AppError('INTERVALSRVCE', error);
+			throw new AppError('[INTERVAL_SERVICE]', error);
 		}
 	}
 
 	public async removeByRemoteId(remoteId: string): Promise<void> {
 		try {
+			if (!remoteId || typeof remoteId !== 'string') {
+				return console.error('WARN[INTERVAL_SERVICE]: No interval remoteId, cannot remove');
+			}
 			await this._intervalDAO.deleteByRemoteId(remoteId);
 		} catch (error) {
-			throw new AppError('INTERVALSRVCE', error);
+			throw new AppError('[INTERVAL_SERVICE]', error);
 		}
 	}
