+++ packages/desktop-libs/src/lib/offline/services/timer.service.ts
@@ -11,6 +11,7 @@ export class TimerService implements ITimerService<TimerTO> {
 		this._timerDAO = new TimerDAO();
 		this._userService = new UserService();
 	}
+
 	public async findLastOne(): Promise<TimerTO> {
 		try {
 			const user = await this._userService.retrieve();
@@ -20,6 +21,7 @@ export class TimerService implements ITimerService<TimerTO> {
 			return null;
 		}
 	}
+
 	public async findLastCapture(): Promise<TimerTO> {
 		try {
 			const user = await this._userService.retrieve();
@@ -29,13 +31,18 @@ export class TimerService implements ITimerService<TimerTO> {
 			return null;
 		}
 	}
+
 	public async update(timer: Partial<Timer>): Promise<void> {
 		try {
+			if (!timer && !timer.id) {
+				return console.error('WARN[TIMER_SERVICE]: No timer data, cannot update');
+			}
 			await this._timerDAO.update(timer.id, timer.toObject());
 		} catch (error) {
-			throw new AppError('TMRSRVCE', error);
+			throw new AppError('[TIMER_SERVICE]', error);
 		}
 	}
+
 	public async findAll(): Promise<TimerTO[]> {
 		try {
 			return await this._timerDAO.findAll();
@@ -44,27 +51,41 @@ export class TimerService implements ITimerService<TimerTO> {
 			return [];
 		}
 	}
+
 	public async findById(timer: Partial<Timer>): Promise<TimerTO> {
 		try {
+			if (!timer && !timer.id) {
+				console.error('WARN[TIMER_SERVICE]: No timer data, cannot find');
+				return null;
+			}
 			return await this._timerDAO.findOneById(timer.id);
 		} catch (error) {
 			console.error(error);
 			return null;
 		}
 	}
+
 	public async remove(timer: Partial<Timer>): Promise<void> {
 		try {
+			if (!timer) {
+				console.error('WARN[TIMER_SERVICE]: No timer data, cannot remove');
+				return null;
+			}
 			await this._timerDAO.delete(timer);
 		} catch (error) {
-			throw new AppError('TMRSRVCE', error);
+			throw new AppError('[TIMER_SERVICE]', error);
 		}
 	}
 
 	public async save(timer: Timer): Promise<void> {
 		try {
+			if (!timer) {
+				console.error('WARN[TIMER_SERVICE]: No timer data, cannot save');
+				return null;
+			}
 			await this._timerDAO.save(timer.toObject());
 		} catch (error) {
-			throw new AppError('TMRSRVCE', error);
+			throw new AppError('[TIMER_SERVICE]', error);
 		}
 	}
